move entity interfaces engine config to separate file
